Self updating now uses dedicated script to replace file itself.
git-svn-id: http://typo3scripts.googlecode.com/svn/branches/dev@31 a8a2147e-0eb0-e98b-0563-54e275c8f842

+++ createBackup.sh
@@ -57,14 +57,17 @@ UPDATE_BASE=http://typo3scripts.googlecode.com/svn/trunk
 
 # Self-update
 runSelfUpdate() {
-  echo -n "Performing self-update..."
+  echo "Performing self-update..."
+  
   
   # Download new version
+  echo -n "Downloading latest version..."
   if ! wget --quiet --output-document="$0.tmp" $UPDATE_BASE/$SELF ; then
     echo "Failed: Error while trying to wget new version!"
     echo "File requested: $UPDATE_BASE/$SELF"
     exit 1
   fi
+  echo "Done."
   
   # Copy over modes from old version
   OCTAL_MODE=$(stat -c '%a' $SELF)
@@ -73,11 +76,20 @@ runSelfUpdate() {
     exit 1
   fi
   
-  # Overwrite old file with new
-  mv "$0.tmp" "$0"
+  # Spawn update script
+  cat > updateScript.sh << EOF
+#!/bin/bash
+# Overwrite old file with new
+if mv "$0.tmp" "$0"; then
+  echo "Done. Update complete."
+  rm \$0
+else
+  echo "Failed!"
+fi
+EOF
   
-  echo "Done"
-  exit 0
+  echo -n "Inserting update process..."
+  exec updateScript.sh
 }
 
 # Read external configuration
@@ -130,12 +142,21 @@ fi
 
 FILE=$BASE-$(date +%Y-%m-%d-%H-%M).tgz
 echo "Creating Typo3 backup '$FILE'..."
+
+# Create database dump
 echo -n "Creating database dump at $BASE/database.sql..."
-mysqldump --host=$HOST --user=$USER --password=$PASS --add-drop-table --add-drop-database --databases $DB > $BASE/database.sql
+if ! mysqldump --host=$HOST --user=$USER --password=$PASS --add-drop-table --add-drop-database --databases $DB > $BASE/database.sql; then
+  echo "Failed!"
+  exit 1
+fi
 echo "Done."
+
+# Create backup archive
 echo -n "Compressing Typo3 installation..."
 tar --create --gzip --file $FILE $BASE > /dev/null
 echo "Done."
+
+# Now that the database dump is packed up, delete it
 echo -n "Deleting database dump..."
 rm --force $BASE/database.sql
 echo "Done!"
